Add security to all process pages to make sure users cannot enter them

+++ removeFromList_Process.php
@@ -1,15 +1,20 @@
 <?php
+if(!empty($_POST)){
+    $query = "DELETE FROM userlist WHERE movieID =" . $_REQUEST['id'] . " AND userID=" . $_SESSION['id'] . ";";
+        $result = $db->query($query);
+        if (!$result) {
+            print "Error - the query could not be executed";
+            $error = mysqli_error($db);
+            print "<p>" . $error . "</p>";
+            exit;
+        }
 
-$query = "DELETE FROM userlist WHERE movieID =" . $_REQUEST['id'] . " AND userID=" . $_SESSION['id'] . ";";
-    $result = $db->query($query);
-    if (!$result) {
-        print "Error - the query could not be executed";
-        $error = mysqli_error($db);
-        print "<p>" . $error . "</p>";
-        exit;
-    }
+        $_SESSION['movieList'] = array_diff( $_SESSION['movieList'], [$_REQUEST['id']] );
 
-    $_SESSION['movieList'] = array_diff( $_SESSION['movieList'], [$_REQUEST['id']] );
-
-    header("Location: ./?action=main.php");
+        header("Location: ./?action=main.php");
+}
+else{
+    print("Access Denied");
+    exit();
+}
 ?>
